scingest/requirements.txt
# can be just plain Pillow
Pillow-SIMD==4.3.0.post0
python-sane==2.8.2
